refactor : Modal파일들 Modal 폴더에서 관리 & 모달 네이밍 수정
