+++ client/src/components/Modal/DeleteMindMapModal.tsx
