+++ client/src/components/Modal/GuestNewMindMapModal.tsx
