+++ client/src/components/Modal/LatestMindMapModal.tsx
