+++ client/src/pages/layout/index.tsx
@@ -4,7 +4,7 @@ import { useSideBar } from "@/store/useSideBar";
 import { ErrorBoundary } from "react-error-boundary";
 import { Outlet } from "react-router-dom";
 import { useConnectionStore } from "@/store/useConnectionStore";
-import OfflineModal from "@/components/OfflineModal";
+import OfflineModal from "@/components/Modal/OfflineModal";
 import useWindowEventListener from "@/hooks/useWindowEventListener";
 import { useState } from "react";
 import NotFound from "@/components/common/NotFound";
